Load kernel at 0x100000 and split bootloader into 2 stages.
We also package our image as a CDROM in ISO format.
The new bootloader:
(1) Stage 1 is in sector 1 of the image and gets loaded by BIOS.
(2) Stage 1 then uses BIOS to read disk, load stage 2 in memory, find the
location and size of the kernel and give that info to stage 2.
(3) Stage 2 reads kernel info, switches to unreal mode, copies kernel in chunks
first to a lower address using BIOS and then copies it to a higher address.
(4) It then loads GDT, switches to 32-bit protected mode and starts executing
the kernel.